requirements.txt
tensorflow-cpu==2.15.1
streamlit>=1.32.0
streamlit-drawable-canvas>=0.9.3
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0